+++ frontend/components/skeletons/chart-skeleton.tsx
@@ -1,0 +1,5 @@
+import { Skeleton } from "../ui/skeleton";
+
+export default function ChartSkeleton() {
+  return <Skeleton className="w-full h-[300px] rounded-lg" />;
+}
